requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.5
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-filter==24.3
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-spectacular==0.28.0
flower==2.0.1
greenlet==3.2.4
humanize==4.13.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
